requirements.txt
python-logstash-async==1.0.0